Refactor fetching from localStorage into custom hook
